+++ ios/BTE/Extensions/Notification+Extensions.swift
@@ -1,0 +1,7 @@
+extension Notification.Name {
+  public static let StorageTestResultsDidChange = Notification.Name(rawValue: "BTESecureStorageTestResultsDidChange")
+  public static let StorageExposureDetectionErrorLocalizedDescriptionDidChange = Notification.Name(rawValue: "BTESecureStorageExposureDetectionErrorLocalizedDescriptionDidChange")
+  public static let DateLastPerformedExposureDetectionDidChange = Notification.Name(rawValue: "BTESecureStorageDateLastPerformedExposureDetectionDidChange")
+  public static let ExposuresDidChange = Notification.Name(rawValue: "BTESecureStorageExposuresDidChange")
+  public static let NextDiagnosisKeyFileIndexDidChange = Notification.Name(rawValue: "BTESecureStorageNextDiagnosisKeyFileIndexDidChange")
+}
