+++ ios/BTE/Extensions/String+Extensions.swift
@@ -4,4 +4,9 @@ extension String {
   static let `default` = ""
   static let notAuthorized = "notAuthorized"
   static let authorized = "authorized"
+  static let keyPathTestResults = "testResults"
+  static let keyPathExposureDetectionErrorLocalizedDescription = "exposureDetectionErrorLocalizedDescription"
+  static let keyPathDateLastPerformedExposureDetection = "dateLastPerformedExposureDetection"
+  static let keyPathExposures = "exposures"
+  static let keyPathNextDiagnosisKeyFileIndex = "nextDiagnosisKeyFileIndex"
 }
